code/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
joblib==0.17.0
google-cloud-automl==2.5.2